Implemented the waypoint view.
added features to MapActivity + saveInstance and Restore instance  and
a BroadcastReceiver that receives a Intent from Tracking Service.

In the Trip activity I added a click handler to the Image View of the Map.
The click on it will start the MapActivity and will render the waypoints.

The BroadcastReceiver will render new Waypoints while tracking is activated.